refactored bits of code
