+++ simbad/util/__init__.py
@@ -4,13 +4,14 @@ __author__ = "Adam Simpkin, Felix Simkovic & Jens Thomas"
 __date__ = "05 May 2017"
 __version__ = "1.0"
 
+import iotbx.pdb
+import iotbx.pdb.amino_acid_codes
 import logging
+import pandas as pd
 import string
 
 from simbad.chemistry import atomic_composition, periodic_table
 
-import iotbx.pdb
-import iotbx.pdb.amino_acid_codes
 
 
 logger = logging.getLogger(__name__)
@@ -72,4 +73,29 @@ def molecular_weight(pdbin):
     mw += hydrogen_atoms * periodic_table['H'].atomic_mass
 
     return mw
+
+
+def summarize_result(results, csv_file=None, columns=None):
+    """Summarize the search results"""
+    
+    # Construct the dataframe for the results
+    kwargs = {}
+    if columns:
+        kwargs["columns"] = columns
+    df = pd.DataFrame([r._as_dict() for r in results], index=[r.pdb_code for r in results], **kwargs)
+
+    # Create a CSV for reading later
+    if csv_file:
+        logger.debug("Storing results in file: %s", csv_file)
+        df.to_csv(csv_file)
+    
+    # Display the table in stdout
+    if df.empty:
+        logger.info("No results found")
+    else:
+        summary_table = "The results for this search are:\n\n%s\n"
+        logger.info(summary_table, df.to_string())
+
+    
+
